Add square generation and deletion to script.js
Added variables and funcions for prompring the user for the amount of
squares, generating squares and deleting squares as well as clearing the
array.

TO DO:
When user enters a number, the numOfSquares needs to get updates and
generateSquares function called again with the new value to generate a
new amount of squares.

Each square needs an onmouseover (or similar?) event listener, which
changes the bg color of the square randomly

Each square needs another value from 1 - 10 that will track the darkness
of the background